+++ src/main.zig
@@ -15,10 +15,9 @@ pub fn main() !void {
         .PC = 0,
         .IR = 0,
         .Flags = 0b0000_0000,
-        .Registers = [_]u16{0} ** 8,
+        .Registers = [_]u8{0} ** 8,
         .WorkingMemory = [_]u16{0} ** 256,
         .Program = instructions,
     };
     std.debug.print("CPU!\n{any}\n", .{z});
-    // std.debug.print("CPU!\n{any}\n", .{cpu});
 }
